Isolate Pandas Dependency
Move all code that depends on pandas to a script, where it can be
invoked when needed by the object storage benchmark but not affect any
of the other benchmarks or our CI infrastructure.
