+++ requirements.txt
@@ -19,4 +19,3 @@ blinker>=1.3
 futures>=3.0.3
 PyYAML==3.11
 pint>=0.7
-pandas
